requirements.txt
# Requirements
numpy
pandas
pillow
opencv-contrib-python #has more modules
python-bsonjs 

#for testing
nose
unittest



